Time: 32 ms (59.12%), Space: 106.8 MB (48.03%) - LeetHub

+++ 2126-destroying-asteroids/2126-destroying-asteroids.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool asteroidsDestroyed(int mass, vector<int>& asteroids) {
         sort(asteroids.begin(),asteroids.end());
-        long long z = mass;
+       double z = mass;
         for(int i = 0 ; i < asteroids.size() ; i++)
         {
             if(z >= asteroids[i])
